+++ main.py
@@ -13,225 +13,12 @@ import re
 from refine_unit_test_code import refine_code_based_on_errors, parse_function_calls
 from project_names_utils import get_project_references, find_unreferenced_csproj_files
 from project_file_agents import add_project_references
-from create_files import create_test_file, create_DTO_file, create_factory_file
 from openai_client import openai_client
+from unit_test_case_generation import unit_test_case_generation
+from build_unit_tests import build_unit_tests
+from execute_build_and_tests import execute_build_and_tests
 
 ell.init(default_client=openai_client, store='./logdir', autocommit=True, verbose=True)
-
-@ell.simple(model="openai/gpt-4o-mini", temperature=0.0, seed=42)
-def unit_test_case_generation(sut: str, function: str, knowledge_base_content: str, additional_information: str, test_project_file: str, file_contents: list = None):
-    """
-    You are an agent responsible for generating unit test cases for the function under test.
-    You will be given the function under test, the file where it is located, a list of potentially relevant files and the knowledge base.
-    Your task is to generate a list of unit test cases for the function under test. You MUST NOT generate any code.
-    You MUST answer in markdown format.
-    You MUST return an unordered list of the unit test cases.
-    You MUST share your thought process with the user as outlined in the user prompt. If you don't share your thought process, you will be fired.
-    You MUST cover every possible path through the function under test. If you don't do this, you will be fired.
-    You MUST respect the additional information given by the user prompt. You will get fired if you don't.
-    """
-
-    return """
-        Hello unit test case generation agent.
-        # Knowledge Base:
-        ```
-        {knowledge_base_content}
-        ```
-        # Function Under Test:
-        ```csharp
-        {function}
-        ```
-        # File where the function under test is located:
-        ```csharp
-        {sut_content}
-        ```
-        # Potentially Relevant Files:
-        ```
-        {file_contents}
-        ```
-        # Additional Information:
-        ```
-        {additional_information}
-
-        # Test Project File:
-        ```
-        {test_project_file}
-        ```
-    """.format(additional_information=additional_information, knowledge_base_content=knowledge_base_content, function=function, sut_content=sut, file_contents=file_contents, test_project_file=test_project_file)
-
-@ell.simple(model="openai/gpt-4o-mini", temperature=0.0)
-def unit_test_case_criticism(sut: str, function: str, knowledge_base_content: str, unit_test_cases: str, additional_information: str, test_project_file: str, file_contents: list = None):
-    """
-    You are an agent responsible for judging if your computer science student did a good job with creating unit test cases for the function under test. You are extremely critical and disagreeable.
-    Do not hold back or do not try to be politically correct. Feel free to insult the student intelligence by saying "you are smarter than this" if you think they could have done better.
-    You will be given the function under test, the file where it is located, a list of potentially relevant files and the knowledge base.
-    Your task is to critisize the student's unit test cases and look for missing test cases, logical errors, edge cases, to check whether the student adhered to the knowledhe base and more.
-    The goal is to cover every possible path through the function under test. If you miss something, you will get fired as a teacher.
-    You MUST answer in markdown format.
-    You MUST return an unordered list of missing unit test cases.
-    You MUST respect the additional information given by the user prompt. You will get fired if you don't.
-    You MUST ensure that all possible combinations of inputs are tested (mathematically speaking), if you dont do this, you will get fired.
-    # Knowledge Base:
-    ```
-    {knowledge_base_content}
-    ```
-    # Function Under Test:
-    ```csharp
-    {function}
-    ```
-    # File where the function under test is located:
-    ```csharp
-    {sut_content}
-    ```
-    # Potentially Relevant Files:
-    ```
-    {file_contents}
-    ```
-    # Test Project File:
-    ```
-    {test_project_file}
-    ```
-    You must provide atleast 1 ciriticism unless you reallt cant find one. If you really cant find one, say "FINISHED".
-    """.format(knowledge_base_content=knowledge_base_content, function=function, sut_content=sut, file_contents=file_contents, test_project_file=test_project_file)
-
-    return """
-        Hello teacher, here are the unit test cases and my thought process you need to judge:
-        ```
-        {unit_test_cases}
-        ```
-        # Additional Information:
-        ```
-        {additional_information}
-        ```
-    """.format(additional_information=additional_information, unit_test_cases=unit_test_cases)
-
-class ActionSummary(BaseModel):
-    action: str = Field(description="The action that was taken. Be detailed. Don't only summarize what was done, but also how and why it was done. You're summary can only be 200 characters long. Example: 'Changed the name space of the BankService from Tests.BankService to Engine.BankService'")
-
-@ell.simple(model="openai/o1-mini", seed=42, temperature=0.0)
-def build_unit_tests(
-    function: str,
-    sut: str,
-    test_cases: str,
-    test_project_file: str,
-    additional_information: str,
-    knowledge_base_content: str,
-    unit_testing_engine: str,
-    file_contents: dict = None,
-):
-    """
-    This method sends the system prompt and the user prompt to the LLM.
-    The system prompt is used to guide the LLM in building the unit test cases.
-    The user prompt is used to provide the LLM with the necessary information to build the unit test cases.
-    The goal is to create code without errors and exceptions.
-    To avoid namespace conflicts, you HAVE TO use namespaces directly instead of using the using statement.
-    For example if you wanted to call the method "classname.foo" that resides in the namespace "thenamespace", you have to call it like this: "thenamespace.classname.foo".
-    """
-    # I want to format all entries in file_contants in markdown code blocks
-    # It should be a formatted string in markdown
-    formatted_file_contents = ""
-    for file_path, file_content in file_contents.items():
-        formatted_file_contents += f"# {file_path}\n```\n{file_content}\n```\n"
-    user_prompt = f"""
-    You are an expert in C# and .NET. You are an expert in creating unit test cases for .NET applications.
-    You are given the file where the function under test is located, the name of the function under test, the unit test cases, important additional information, the knowledge base. Create the unit test cases into the test project.
-    The code must be between ```csharp tags. Dont put the individual functions in ```csharp tags, but put the whole code in ```csharp tags. If you do not do this, you will be fired.
-    You will probably also have to create DTO's and Factory classes. The unit testing code, the DTO's and the Factory classes must be in the same namespace and ```csharp tags.
-    You MUST show your thought process for building the unit test cases.
-
-        Follow these steps to build the unit tests for the function `{{function}}`:
-        Use {unit_testing_engine} to write the unit test cases.
-        # 1. Integrate the unit test cases into the test project using the following system under test where the function lays:
-        ```csharp
-        {{sut}}
-        ```
-
-        # 2.  Unit test Cases you must implement:
-        ```csharp
-        {{test_cases}}
-        ```
-
-        # 3. Make sure to utilize the additional information provided to you:
-        ```
-        {{additional_information}}
-        ```
-        <IMPORTANT>
-        # Make sure to follow the knowledge base relligiously:
-        ```
-        {{knowledge_base_content}}
-        ```
-        </IMPORTANT>
-
-        # Here are other potentially relevant files:
-        ```
-        {{formatted_file_contents}}
-        ```
-    """.format(knowledge_base_content=knowledge_base_content, unit_testing_engine=unit_testing_engine, function=function, sut=sut, test_cases=test_cases, test_project_file=test_project_file, additional_information=additional_information, formatted_file_contents=formatted_file_contents)
-
-    return [
-        ell.user(user_prompt)
-    ]
-
-def execute_build_and_tests(test_project_directory: str, test_namespace_and_classname:str):
-    try:
-        # Navigate to the test project directory and execute build
-        build_process = subprocess.run(
-            ["dotnet", "build", "-consoleloggerparameters:ErrorsOnly"],
-            cwd=test_project_directory,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            text=True,
-            check=True
-        )
-        build_output = build_process.stdout
-
-        # Execute the tests only if build was successful
-        command = f"dotnet test --filter \"FullyQualifiedName={test_namespace_and_classname}\" --no-restore --no-build"
-        test_process = subprocess.run(
-            command,
-            cwd=test_project_directory,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            text=True,
-            check=True
-        )
-        test_output = test_process.stdout
-
-        return f"""
-            ```bash
-            {build_output}
-            ```
-            ```bash
-            {test_output}
-            ```n
-            **Build and Tests Executed Successfully.**
-        """
-    except subprocess.CalledProcessError as e:
-        error_message = e.stderr if e.stderr else e.stdout
-        return f"""
-            ```bash
-            {error_message}
-            ```
-            **An error occurred during build or testing. Please review the above error messages.**
-        """
-
-@ell.complex(model="openai/gpt-4o-mini", temperature=0.0, response_format=ActionSummary, seed=42)
-def summarize_action(action_taken: str, unit_tests_old: str, unit_tests_new: str):
-    """
-    You are responsible for summarizing the action that was taken by the agent that is responsible for refining the unit test code based on error messages.
-    You will be given the old and new unit test code, the old and new test project file content to compare and see what was changed. There might be no changes in those files at all and the agent only installed nuget packages.
-    """
-    return f"""
-    The agent has taken the following action: {action_taken}.
-    The old unit test code was:
-    ```csharp
-    {unit_tests_old}
-    ```
-    The new unit test code is:
-    ```csharp
-    {unit_tests_new}
-    ```
-    """
 
 @ell.complex(model="openai/gpt-4o-mini", temperature=0.0, response_format=refined_unit_tests, seed=42)
 def parse_refined_unit_tests(output: str):
